Remove constant timeslice script
Script doesn't seem to be particularly useful.

If cache is ON on the environment, the only thing we will be testing is
network speed and we should be able to get transfer time from standard
k6 metrics (http_req_receiving).

If cache is OFF on the environment then script will behave similar to
the random slice script.