Remove random horizon scenario
This is the most harsh setup we have as it requires download of many
blobs (data is located way too far from each other). It has nothing to
do with real user pattern where horizons are continuous.
So testing for this case actually causes more problems and confusion
than not.